refactor: :recycle: svg는 수동적으로 props를 받기만 하기
