+++ app/src/assets/AppTitleSvg.tsx
@@ -1,22 +1,19 @@
-import { AppLogoSize } from '@/utils/types/AppLogoSize';
+import { useTheme } from '@emotion/react';
+import { SVGProps } from 'react';
 
-type AppTitleSvgProps = {
-  size: AppLogoSize;
-};
+export const AppTitleSvg = (props: SVGProps<SVGSVGElement>) => {
+  const theme = useTheme();
 
-export const AppTitleSvg = ({ size }: AppTitleSvgProps) => {
-  const rate = size === 'sm' ? 0.75 : 1;
   return (
     <svg
-      width={135 * rate}
-      height={25 * rate}
       viewBox="0 0 135 25"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      {...props}
     >
       <path
         d="M12.1859 19.322H6.43986C4.2412 19.322 2.65453 18.8687 1.67986 17.962C0.727863 17.0327 0.251863 15.6047 0.251863 13.678V1.744C0.251863 1.31333 0.34253 1.00733 0.523863 0.825999C0.727863 0.621999 1.0452 0.519999 1.47586 0.519999H6.74586C7.17653 0.519999 7.48253 0.621999 7.66386 0.825999C7.86786 1.00733 7.96986 1.31333 7.96986 1.744V11.06C7.96986 11.4227 8.06053 11.706 8.24186 11.91C8.44586 12.114 8.70653 12.216 9.02386 12.216H12.1859V1.744C12.1859 1.31333 12.2765 1.00733 12.4579 0.825999C12.6619 0.621999 12.9792 0.519999 13.4099 0.519999H18.6459C19.0765 0.519999 19.3825 0.621999 19.5639 0.825999C19.7679 1.00733 19.8699 1.31333 19.8699 1.744V23.776C19.8699 24.2067 19.7679 24.524 19.5639 24.728C19.3825 24.9093 19.0765 25 18.6459 25H13.4099C12.9792 25 12.6619 24.9093 12.4579 24.728C12.2765 24.524 12.1859 24.2067 12.1859 23.776V19.322ZM41.6633 25H25.4793C25.0487 25 24.7313 24.9093 24.5273 24.728C24.346 24.524 24.2553 24.2067 24.2553 23.776V17.928C24.2553 16.6133 24.38 15.5367 24.6293 14.698C24.9013 13.8367 25.2753 13.134 25.7513 12.59C26.2273 12.046 26.794 11.6153 27.4513 11.298C28.1087 10.9807 28.8227 10.7087 29.5933 10.482L34.1493 9.054C34.512 8.96333 34.75 8.82733 34.8633 8.646C34.9993 8.46467 35.0673 8.19267 35.0673 7.83C35.0673 7.49 34.9993 7.26333 34.8633 7.15C34.7273 7.014 34.4893 6.946 34.1493 6.946H26.7713C26.3407 6.946 26.0233 6.85533 25.8193 6.674C25.638 6.47 25.5473 6.15267 25.5473 5.722V1.744C25.5473 1.31333 25.638 1.00733 25.8193 0.825999C26.0233 0.621999 26.3407 0.519999 26.7713 0.519999H36.8013C38.728 0.519999 40.2013 1.00733 41.2213 1.982C42.264 2.934 42.7853 4.328 42.7853 6.164V9.36C42.7853 11.06 42.3093 12.3633 41.3573 13.27C40.4053 14.1767 39.1133 14.8793 37.4813 15.378L32.8573 16.84C32.5173 16.9307 32.2793 17.0327 32.1433 17.146C32.03 17.2593 31.9733 17.4973 31.9733 17.86V18.574H41.6633C42.094 18.574 42.4 18.676 42.5813 18.88C42.7853 19.0613 42.8873 19.3673 42.8873 19.798V23.776C42.8873 24.2067 42.7853 24.524 42.5813 24.728C42.4 24.9093 42.094 25 41.6633 25ZM58.9922 25H48.2482C47.8175 25 47.5002 24.9093 47.2962 24.728C47.1148 24.524 47.0242 24.2067 47.0242 23.776V19.798C47.0242 19.3673 47.1148 19.0613 47.2962 18.88C47.5002 18.676 47.8175 18.574 48.2482 18.574H57.3942C57.6888 18.574 57.9042 18.4833 58.0402 18.302C58.1988 18.098 58.2782 17.86 58.2782 17.588C58.2782 17.2027 58.1988 16.9193 58.0402 16.738C57.9042 16.5567 57.6888 16.4547 57.3942 16.432L52.2942 15.718C50.7075 15.4913 49.4268 14.97 48.4522 14.154C47.5002 13.338 47.0242 11.978 47.0242 10.074V6.504C47.0242 4.57733 47.6475 3.104 48.8942 2.084C50.1408 1.04133 51.8635 0.519999 54.0622 0.519999H63.5822C64.0128 0.519999 64.3188 0.621999 64.5002 0.825999C64.7042 1.00733 64.8062 1.31333 64.8062 1.744V5.79C64.8062 6.22067 64.7042 6.538 64.5002 6.742C64.3188 6.92333 64.0128 7.014 63.5822 7.014H55.6602C55.3655 7.014 55.1388 7.116 54.9802 7.32C54.8215 7.50133 54.7422 7.73933 54.7422 8.034C54.7422 8.32867 54.8215 8.578 54.9802 8.782C55.1388 8.986 55.3655 9.09933 55.6602 9.122L60.7262 9.802C62.3128 10.0287 63.5935 10.55 64.5682 11.366C65.5428 12.182 66.0302 13.542 66.0302 15.446V19.016C66.0302 20.9427 65.3955 22.4273 64.1262 23.47C62.8795 24.49 61.1682 25 58.9922 25ZM81.6574 25H76.1494C75.7188 25 75.4014 24.9093 75.1974 24.728C75.0161 24.524 74.9254 24.2067 74.9254 23.776V7.082H69.9274C69.4968 7.082 69.1794 6.99133 68.9754 6.81C68.7941 6.606 68.7034 6.28867 68.7034 5.858V1.744C68.7034 1.31333 68.7941 1.00733 68.9754 0.825999C69.1794 0.621999 69.4968 0.519999 69.9274 0.519999H87.8794C88.3101 0.519999 88.6161 0.621999 88.7974 0.825999C89.0014 1.00733 89.1034 1.31333 89.1034 1.744V5.858C89.1034 6.28867 89.0014 6.606 88.7974 6.81C88.6161 6.99133 88.3101 7.082 87.8794 7.082H82.8814V23.776C82.8814 24.2067 82.7794 24.524 82.5754 24.728C82.3941 24.9093 82.0881 25 81.6574 25ZM100.603 7.66L99.0046 12.828H104.343L102.745 7.66C102.654 7.456 102.552 7.30867 102.439 7.218C102.348 7.12733 102.235 7.082 102.099 7.082H101.249C101.113 7.082 100.988 7.12733 100.875 7.218C100.784 7.30867 100.693 7.456 100.603 7.66ZM98.6646 23.776C98.6646 24.2067 98.5626 24.524 98.3586 24.728C98.1773 24.9093 97.8713 25 97.4406 25H92.4426C92.0119 25 91.6946 24.9093 91.4906 24.728C91.3093 24.524 91.2186 24.2067 91.2186 23.776V14.8C91.2186 14.12 91.3206 13.338 91.5246 12.454C91.7513 11.57 92.0913 10.5387 92.5446 9.36L95.5026 1.642C95.6386 1.25666 95.8426 0.973332 96.1146 0.792C96.4093 0.610666 96.7833 0.519999 97.2366 0.519999H106.383C106.813 0.519999 107.165 0.610666 107.437 0.792C107.731 0.973332 107.947 1.25666 108.083 1.642L111.041 9.36C111.494 10.5387 111.823 11.57 112.027 12.454C112.253 13.338 112.367 14.12 112.367 14.8V23.776C112.367 24.2067 112.265 24.524 112.061 24.728C111.879 24.9093 111.573 25 111.143 25H106.009C105.555 25 105.215 24.9093 104.989 24.728C104.785 24.524 104.683 24.2067 104.683 23.776V19.016H98.6646V23.776ZM127.443 25H121.935C121.505 25 121.187 24.9093 120.983 24.728C120.802 24.524 120.711 24.2067 120.711 23.776V7.082H115.713C115.283 7.082 114.965 6.99133 114.761 6.81C114.58 6.606 114.489 6.28867 114.489 5.858V1.744C114.489 1.31333 114.58 1.00733 114.761 0.825999C114.965 0.621999 115.283 0.519999 115.713 0.519999H133.665C134.096 0.519999 134.402 0.621999 134.583 0.825999C134.787 1.00733 134.889 1.31333 134.889 1.744V5.858C134.889 6.28867 134.787 6.606 134.583 6.81C134.402 6.99133 134.096 7.082 133.665 7.082H128.667V23.776C128.667 24.2067 128.565 24.524 128.361 24.728C128.18 24.9093 127.874 25 127.443 25Z"
-        fill="#161616"
+        fill={theme.colors.mono.black}
       />
     </svg>
   );
