+++ app/src/components/elements/AppLogoTitleButton.tsx
@@ -1,4 +1,4 @@
-import { AppLogoTitle } from '@/assets/AppLogoTitle';
+import { AppLogoTitle } from '@/components/elements/AppLogoTitle';
 import { useNavigate } from 'react-router-dom';
 import { Button } from '@/components/common';
