+++ app/src/utils/atoms/isAuthenticatedAtom.ts
@@ -1,3 +1,3 @@
 import { atom } from 'jotai';
 
-export const isAuthenticatedAtom = atom(true);
+export const isAuthenticatedAtom = atom(false);
